Add CI (#3)
* Do a partial clone of linux

This saves a lot of disk space. The git dir for the partial clone after
all relevant versions have been checked out is 1.7GB.

* Do a sparse checkout of linux

This saves a lot of disk space. The git dir for the partial clone after all relevant versions have been checked out is <800MB.

* Fix typo

* Fix checkout on CI

* Add CI

* Make generation deterministic

Directory iteration order is non-deterministic. In practice it depends
on the exact order in which dir entries have been created. Sort the
output of fs::read_dir to ensure that generation is deterministic across
systems.

* Check for linux/.git

* Build generator in release mode

* Switch to cargo check