Simplify the bindings by using only a single kernel version.
Originally, the idea of having multiple versions was to attempt to
support writing code that was strictly compatible with old versions.
However, my experience with linux-raw-sys so far as been that that's
not as useful as it seemed like it would be, and versioning all the
imports means having multiple versions of all the types, which gets
awkward.

So instead, switch to having just a single Linux version, 5.11 for
now, which is mostly a superset of the other versions.

This also helps reduce compile times, as there aren't multiple
versions of everything to parse.