Add tests to ensure that the C type definitions match libc. (#27)
* Add tests to ensure that the C type definitions match libc.

In no-std mode, we don't depend on libc. But, we still want to be sure
that our types match those of libc. This adds a libc dev-dependency and
checks that the types are the same in cargo-test.

* Use `cargo check --tests` instead of `cargo test`.

We only have static-assertion tests, and `cargo check --tests` is enough
for those, and doesn't require installing cross-compilers.