Add definition of `c_char` for SPARC
SPARC 32 bit is a tier 3 platform in Rust, though rustix checks against
it. A proposed update to 0.6.4 for this library degraded those checks as
SPARC 32 bit lacks a definition for `c_char`.

The best evidence I could find as to the signedness of SPARC 32 bit is
the blog post linked below, though I have not replicated their findings.

https://github.com/bytecodealliance/rustix/actions/runs/8955815255/job/24596945878?pr=1056
https://trofi.github.io/posts/203-signed-char-or-unsigned-char.html