Make ioctl generation use correct version of Linux headers
Ioctls now use the same kernel version that's used in `gen/src/main.rs` for
generating other bindings. They are no longer based on any system-wide headers
which happen to exist in the system include directories.

This removes the `REISERFS_IOC_UNPACK` ioctl because reiserfs was removed from
the kernel and its userspace headers in Linux 6.13.